training: replace action cap with max_hands for exact 1-hand games
The action cap (175) was a crude proxy for stopping after one hand — it
happened to cut off near the 1st hand boundary, but 1.6% of games hit the
cap mid-hand (scores=[0,0], a draw) and a trained policy could shift the
action distribution enough to throw off the tuning.

Add `max_hands` to the Rust Pool: after settle_hand banks the scores,
if `hands >= max_hands` the game stops with its banked scores and
`unfinished=false`. A hand is guaranteed to finish in finite time (the
stock depletes), so no action cap is needed. The training path uses
`max_hands=1`; the action cap is removed from the league, train, bench,
and shards paths. The Pool still supports the action cap for other
consumers (PG trainer, evaluator).

+++ training/python/canastra_train/elo.py
@@ -29,7 +29,7 @@ class EloTracker:
         self.base = base
 
     def expected(self, a: int, b: int) -> float:
-        return 1.0 / (1.0 + 10.0 ** ((self.ratings[b] - self.ratings[a]) / 400.0))
+        return float(1.0 / (1.0 + 10.0 ** ((self.ratings[b] - self.ratings[a]) / 400.0)))
 
     def update(self, a: int, b: int, result: float) -> None:
         exp_a = self.expected(a, b)
